break up utils file
